Refactor Processing
Moving files around to seperate the loading of a memory and checking parts.
