+++ src/main/java/processing/load/MemoryCreate.java
@@ -1,30 +1,16 @@
-package processing;
+package processing.load;
 
-import com.drew.imaging.ImageMetadataReader;
-import com.drew.imaging.ImageProcessingException;
-import com.drew.metadata.Metadata;
-import com.drew.metadata.exif.ExifIFD0Directory;
-import com.drew.metadata.exif.ExifSubIFDDirectory;
 import obj.Memory;
 import obj.UserInput;
+import processing.Shared;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.File;
-import java.io.IOException;
 import java.util.List;
 
 class MemoryCreate implements Runnable
 {
-    private static final String[] PICTURE_EXTENSIONS = new String[]{
-            "gif", "png", "bmp", "jpg", "jpeg", "heic"
-    };
-    private static final String[] PICTURE_RAW_EXTENSIONS = new String[]{
-            "raw", "cr2", "tiff", "tif"
-    };
-    private static final String[] VIDEO_EXTENSIONS = new String[]{
-            "mp4", "mov", "mp4", "mov", "avi", "vlc", "wmv"
-    };
     private final File file;
     private final List<Memory> memories;
     private final UserInput userIo;
@@ -38,7 +24,7 @@ class MemoryCreate implements Runnable
 
     private static boolean isVideo(final File file)
     {
-        for (final String x : VIDEO_EXTENSIONS)
+        for (final String x : Shared.VIDEO_EXTENSIONS)
         {
             if (file.getName().toLowerCase().endsWith(x))
             {
@@ -50,7 +36,7 @@ class MemoryCreate implements Runnable
 
     private static boolean isPicture(final File file)
     {
-        for (final String x : PICTURE_EXTENSIONS)
+        for (final String x : Shared.PICTURE_EXTENSIONS)
         {
             if (file.getName().toLowerCase().endsWith(x))
             {
@@ -62,7 +48,7 @@ class MemoryCreate implements Runnable
 
     private static boolean isRaw(final File file)
     {
-        for (final String x : PICTURE_RAW_EXTENSIONS)
+        for (final String x : Shared.PICTURE_RAW_EXTENSIONS)
         {
             if (file.getName().toLowerCase().endsWith(x))
             {
@@ -103,24 +89,8 @@ class MemoryCreate implements Runnable
                     }
                 } catch (final Exception ioe)
                 {
-                    try
-                    {
-                        final Metadata metadata = ImageMetadataReader.readMetadata(file);
-                        memory.setMetadata(metadata);
-
-                        final ExifIFD0Directory exifIFD0Directory = metadata.getFirstDirectoryOfType(ExifIFD0Directory.class);
-                        memory.setWidth(Integer.parseInt(exifIFD0Directory.getString(ExifIFD0Directory.TAG_IMAGE_WIDTH)));
-                        memory.setHeight(Integer.parseInt(exifIFD0Directory.getString(ExifIFD0Directory.TAG_IMAGE_HEIGHT)));
-
-                        final ExifSubIFDDirectory exifSubIFDDirectory = metadata.getFirstDirectoryOfType(ExifSubIFDDirectory.class);
-                        memory.setDate(exifSubIFDDirectory.getString(ExifSubIFDDirectory.TAG_DATETIME_ORIGINAL));
-                        memory.setMetaDataLoaded(true);
-
-                    } catch (final ImageProcessingException | IOException | NullPointerException ipe)
-                    {
-                        System.out.println("Error loading memory: " + file.getName());
-                        ipe.printStackTrace();
-                    }
+                    System.out.println("Error loading memory: " + file.getPath());
+                    ioe.printStackTrace();
                 }
 
                 this.memories.add(memory);
